+++ packages/plugin-rsc/e2e/starter.test.ts
@@ -1,94 +1,55 @@
 import { expect, test } from '@playwright/test'
-import { type Fixture, useFixture } from './fixture'
-import { expectNoReload, testNoJs, waitForHydration } from './helper'
+import { useFixture } from './fixture'
+import { defineStarterTest } from './starter'
+import { waitForHydration } from './helper'
 
 test.describe('dev-default', () => {
   const f = useFixture({ root: 'examples/starter', mode: 'dev' })
-  defineTest(f)
+  defineStarterTest(f)
 })
 
 test.describe('build-default', () => {
   const f = useFixture({ root: 'examples/starter', mode: 'build' })
-  defineTest(f)
+  defineStarterTest(f)
 })
 
-test.describe('dev-cloudflare', () => {
-  const f = useFixture({ root: 'examples/starter-cf-single', mode: 'dev' })
-  defineTest(f)
+test.describe('dev-production', () => {
+  const f = useFixture({
+    root: 'examples/starter',
+    mode: 'dev',
+    cliOptions: {
+      env: { NODE_ENV: 'production' },
+    },
+  })
+  defineStarterTest(f, 'dev-production')
+
+  test('verify production', async ({ page }) => {
+    await page.goto(f.url())
+    await waitForHydration(page)
+    const res = await page.request.get(f.url('src/client.tsx'))
+    expect(await res.text()).not.toContain('jsxDEV')
+  })
 })
 
-test.describe('build-cloudflare', () => {
-  const f = useFixture({ root: 'examples/starter-cf-single', mode: 'build' })
-  defineTest(f)
+test.describe('build-development', () => {
+  const f = useFixture({
+    root: 'examples/starter',
+    mode: 'build',
+    cliOptions: {
+      env: { NODE_ENV: 'development' },
+    },
+  })
+  defineStarterTest(f)
+
+  test('verify development', async ({ page }) => {
+    let output!: string
+    page.on('response', async (response) => {
+      if (response.url().match(/\/assets\/client-[\w-]+\.js$/)) {
+        output = await response.text()
+      }
+    })
+    await page.goto(f.url())
+    await waitForHydration(page)
+    expect(output).toContain('jsxDEV')
+  })
 })
-
-function defineTest(f: Fixture) {
-  test('basic', async ({ page }) => {
-    await page.goto(f.url())
-    await waitForHydration(page)
-  })
-
-  test('client component', async ({ page }) => {
-    await page.goto(f.url())
-    await waitForHydration(page)
-    await page.getByRole('button', { name: 'Client Counter: 0' }).click()
-    await expect(
-      page.getByRole('button', { name: 'Client Counter: 1' }),
-    ).toBeVisible()
-  })
-
-  test('server action @js', async ({ page }) => {
-    await page.goto(f.url())
-    await waitForHydration(page)
-    await using _ = await expectNoReload(page)
-    await page.getByRole('button', { name: 'Server Counter: 0' }).click()
-    await expect(
-      page.getByRole('button', { name: 'Server Counter: 1' }),
-    ).toBeVisible()
-  })
-
-  testNoJs('server action @nojs', async ({ page }) => {
-    await page.goto(f.url())
-    await page.getByRole('button', { name: 'Server Counter: 1' }).click()
-    await expect(
-      page.getByRole('button', { name: 'Server Counter: 2' }),
-    ).toBeVisible()
-  })
-
-  test('client hmr', async ({ page }) => {
-    test.skip(f.mode === 'build')
-
-    await page.goto(f.url())
-    await waitForHydration(page)
-    await page.getByRole('button', { name: 'Client Counter: 0' }).click()
-    await expect(
-      page.getByRole('button', { name: 'Client Counter: 1' }),
-    ).toBeVisible()
-
-    const editor = f.createEditor(`src/client.tsx`)
-    editor.edit((s) => s.replace('Client Counter', 'Client [edit] Counter'))
-    await expect(
-      page.getByRole('button', { name: 'Client [edit] Counter: 1' }),
-    ).toBeVisible()
-
-    // check next ssr is also updated
-    const res = await page.goto(f.url())
-    expect(await res?.text()).toContain('Client [edit] Counter')
-    await waitForHydration(page)
-    editor.reset()
-    await page.getByRole('button', { name: 'Client Counter: 0' }).click()
-  })
-
-  test('image assets', async ({ page }) => {
-    await page.goto(f.url())
-    await waitForHydration(page)
-    await expect(page.getByAltText('Vite logo')).not.toHaveJSProperty(
-      'naturalWidth',
-      0,
-    )
-    await expect(page.getByAltText('React logo')).not.toHaveJSProperty(
-      'naturalWidth',
-      0,
-    )
-  })
-}