+++ packages/plugin-rsc/examples/basic/src/routes/style-server/server.tsx
@@ -8,6 +8,12 @@ export function TestStyleServer() {
       <div data-testid="css-module-server" className={styles.server}>
         test-css-module-server
       </div>
+      <link
+        rel="stylesheet"
+        href="/test.css"
+        precedence="test-style-server-manual"
+      />
+      <div className="test-style-server-manual">test-style-server-manual</div>
     </>
   )
 }