+++ packages/plugin-rsc/examples/basic/vite.config.ts
@@ -6,35 +6,16 @@ import { type Plugin, defineConfig, normalizePath, parseAstAsync } from 'vite'
 import inspect from 'vite-plugin-inspect'
 import path from 'node:path'
 
-// log unhandled rejection to debug e2e failures
-if (!(globalThis as any).__debugHandlerRegisterd) {
-  process.on('uncaughtException', (err) => {
-    console.error('⚠️⚠️⚠️ uncaughtException ⚠️⚠️⚠️', err)
-  })
-  process.on('unhandledRejection', (err) => {
-    console.error('⚠️⚠️⚠️ unhandledRejection ⚠️⚠️⚠️', err)
-  })
-  ;(globalThis as any).__debugHandlerRegisterd = true
-}
-
 export default defineConfig({
-  base: process.env.TEST_BASE ? '/custom-base/' : undefined,
   clearScreen: false,
   plugins: [
     tailwindcss(),
-    process.env.TEST_REACT_COMPILER
-      ? react({
-          babel: { plugins: ['babel-plugin-react-compiler'] },
-        }).map((p) => ({
-          ...p,
-          applyToEnvironment: (e) => e.name === 'client',
-        }))
-      : react(),
+    react(),
     vitePluginUseCache(),
     rsc({
       entries: {
-        client: './src/client.tsx',
-        ssr: './src/server.ssr.tsx',
+        client: './src/framework/entry.browser.tsx',
+        ssr: './src/framework/entry.ssr.tsx',
         rsc: './src/server.tsx',
       },
       // disable auto css injection to manually test `loadCss` feature.
@@ -79,11 +60,15 @@ export default defineConfig({
         assert(typeof viteManifest.source === 'string')
         if (this.environment.name === 'rsc') {
           assert(viteManifest.source.includes('src/server.tsx'))
-          assert(!viteManifest.source.includes('src/client.tsx'))
+          assert(
+            !viteManifest.source.includes('src/framework/entry.browser.tsx'),
+          )
         }
         if (this.environment.name === 'client') {
           assert(!viteManifest.source.includes('src/server.tsx'))
-          assert(viteManifest.source.includes('src/client.tsx'))
+          assert(
+            viteManifest.source.includes('src/framework/entry.browser.tsx'),
+          )
         }
       },
     },
@@ -138,6 +123,8 @@ export default { fetch: handler };
             source: `\
 /favicon.ico
   Cache-Control: public, max-age=3600, s-maxage=3600
+/test.css
+  Cache-Control: public, max-age=3600, s-maxage=3600
 /assets/*
   Cache-Control: public, max-age=31536000, immutable
 `,
@@ -150,12 +137,25 @@ export default { fetch: handler };
     minify: false,
     manifest: true,
   },
-  optimizeDeps: {
-    exclude: [
-      '@vitejs/test-dep-client-in-server/client',
-      '@vitejs/test-dep-client-in-server2/client',
-      '@vitejs/test-dep-server-in-client/client',
-    ],
+  environments: {
+    client: {
+      optimizeDeps: {
+        entries: [
+          './src/routes/**/client.tsx',
+          './src/framework/entry.browser.tsx',
+        ],
+        exclude: [
+          '@vitejs/test-dep-client-in-server/client',
+          '@vitejs/test-dep-client-in-server2/client',
+          '@vitejs/test-dep-server-in-client/client',
+        ],
+      },
+    },
+    ssr: {
+      optimizeDeps: {
+        include: ['@vitejs/test-dep-transitive-cjs > @vitejs/test-dep-cjs'],
+      },
+    },
   },
 }) as any
 
@@ -166,6 +166,7 @@ function vitePluginUseCache(): Plugin[] {
       async transform(code) {
         if (!code.includes('use cache')) return
         const ast = await parseAstAsync(code)
+        // @ts-ignore for rolldown-vite ci estree/oxc mismatch
         const result = transformHoistInlineDirective(code, ast, {
           runtime: (value) => `__vite_rsc_cache(${value})`,
           directive: 'use cache',
@@ -174,7 +175,7 @@ function vitePluginUseCache(): Plugin[] {
         })
         if (!result.output.hasChanged()) return
         result.output.prepend(
-          `import __vite_rsc_cache from "/src/use-cache-runtime";`,
+          `import __vite_rsc_cache from "/src/framework/use-cache-runtime";`,
         )
         return {
           code: result.output.toString(),